+++ packages/stepper/src/types.ts
@@ -32,8 +32,23 @@ export type SingleStepStates =
   | StepStates.Upcoming;
 
 export interface StepperProps {
+  /**
+   * The index of the step that should be marked as current. (zero-indexed)
+   *
+   * * NOTE: Validations for this prop's value in relation to `completedStepsShown` and `maxDisplayedSteps` are not implemented yet.
+   */
   currentStep: number;
+  /**
+   * Maximum number of steps displayed in the stepper. Includes the ellipses steps.
+   *
+   * * NOTE: Validations for this prop's value in relation to `completedStepsShown` and `currentStep` are not implemented yet.
+   */
   maxDisplayedSteps?: number;
+  /**
+   * Number of completed steps shown before the upcoming steps are displayed. Includes the ellipses step for prior steps.
+   *
+   * * NOTE: Validations for this prop's value in relation to `maxDisplayedSteps` and `currentStep` are not implemented yet.
+   */
   completedStepsShown?: number;
   darkMode?: boolean;
   className?: string;