+++ packages/table/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @leafygreen-ui/table
 
+## 4.1.1
+
+### Patch Changes
+
+- 96d1ff9c: Updates to propTypes, TSDocs, and Storybook controls
+- Updated dependencies [1356d50d]
+- Updated dependencies [6a89bc29]
+- Updated dependencies [fd2f6de0]
+- Updated dependencies [6792bc44]
+- Updated dependencies [96d1ff9c]
+- Updated dependencies [422dbfcd]
+- Updated dependencies [9ff90d4b]
+  - @leafygreen-ui/checkbox@8.0.3
+  - @leafygreen-ui/palette@3.4.0
+  - @leafygreen-ui/hooks@7.3.0
+  - @leafygreen-ui/icon-button@11.0.2
+  - @leafygreen-ui/tokens@1.3.1
+  - @leafygreen-ui/lib@9.3.0
+
 ## 4.1.0
 
 ### Minor Changes